+++ core/web/Request.php
@@ -86,6 +86,6 @@ class Request extends \yii\web\Request
      */
     public function getClientLanguage($defaultValue)
     {
-        return (isset($_SERVER['HTTP_ACCEPT_LANGUAGE'])) ? substr($_SERVER['HTTP_ACCEPT_LANGUAGE'], 0, 2) : $defaultValue;
+        return isset($_SERVER['HTTP_ACCEPT_LANGUAGE']) ? substr($_SERVER['HTTP_ACCEPT_LANGUAGE'], 0, 2) : $defaultValue;
     }
 }